Support for ESP32 is added. It compiles. But yet to be tested properly.

+++ src/utils/Url/Url.h
@@ -14,7 +14,7 @@ class Url {
         String afterPath;
         String query;
         String fragment;
-        uint port;
+        uint16_t port;
         Url();
 };
